FS-2496: Put SysTest forked processes on their own log. Confirm DELETEs and turn down boto logging. Clean up restart test and make sure we hit each service.

+++ source/net-service/svc_request/SvcRequest.cpp
@@ -957,7 +957,7 @@ std::string MultiPrimarySvcRequest::logString()
 {
     std::stringstream oss;
     logSvcReqCommon_(oss, "MultiPrimarySvcRequest");
-    oss << " primaries cnt: " << primariesCnt_;
+    oss << " primaries cnt: " << static_cast<uint32_t>(primariesCnt_);
     return oss.str();
 }
